# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Availability Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  statusNote?: string | null;
  available: boolean;
  availabilityUpdatedAt: string;
  /** @nullable */
  avatarColor?: string | null;
  createdAt: string;
}

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  department?: string;
  statusNote?: string;
  available?: boolean;
  avatarColor?: string;
}

export interface AvailabilityUpdate {
  available: boolean;
  /** @nullable */
  statusNote?: string | null;
}

export interface TeamStats {
  total: number;
  available: number;
  unavailable: number;
}