# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Availability Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: Team member operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a team member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team member found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /users/{id}/availability:
    patch:
      operationId: updateAvailability
      tags: [users]
      summary: Update availability for a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityUpdate"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found

  /stats:
    get:
      operationId: getStats
      tags: [users]
      summary: Get team availability stats
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    TeamMember:
      type: object
      required: [id, name, role, available, availabilityUpdatedAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        statusNote:
          type: ["string", "null"]
        available:
          type: boolean
        availabilityUpdatedAt:
          type: string
        avatarColor:
          type: ["string", "null"]
        createdAt:
          type: string

    TeamMemberInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        department:
          type: string
        statusNote:
          type: string
        available:
          type: boolean
        avatarColor:
          type: string

    AvailabilityUpdate:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        statusNote:
          type: ["string", "null"]

    TeamStats:
      type: object
      required: [total, available, unavailable]
      properties:
        total:
          type: integer
        available:
          type: integer
        unavailable:
          type: integer