# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Availability Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  statusNote?: string | null;
  available: boolean;
  availabilityUpdatedAt: string;
  /** @nullable */
  avatarColor?: string | null;
  createdAt: string;
}