# lib/api-zod/src/generated/types/teamMemberInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Availability Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  department?: string;
  statusNote?: string;
  available?: boolean;
  avatarColor?: string;
}