# lib/api-zod/src/generated/types/teamStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Availability Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamStats {
  total: number;
  available: number;
  unavailable: number;
}